backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
certifi==2023.7.22
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
apscheduler==3.10.4